+++ test/not-sub-exception.js
@@ -3,17 +3,11 @@ const path = require('path')
 const tap = require('tap')
 const clearFolder = require('../cli.js')
 
-const folderErrObj = function ( folder ) {
-    return {
-        message: `clear-folder operates on subfolders of current directory; check ${folder}`
-    }
-}
-
 tap.test(`a message is logged and the operation is aborted
         when the current directory is targeted using '/'`, function ( t ) {
     const folder = '/'
-    const throwing = function () { clearFolder( [folder] ) }
-    t.throws( throwing, folderErrObj( folder ) )
+    const returnValue = clearFolder( [folder] )
+    t.ok( returnValue === -1, `no operation` )
 
     t.end()
 })
@@ -21,8 +15,8 @@ tap.test(`a message is logged and the operation is aborted
 tap.test(`a message is logged and the operation is aborted
         when the current directory is targeted using '.'`, function ( t ) {
     const folder = '.'
-    const throwing = function () { clearFolder( [folder] ) }
-    t.throws( throwing, folderErrObj( folder ) )
+    const returnValue = clearFolder( [folder] )
+    t.ok( returnValue === -1, `no operation` )
 
     t.end()
 })
@@ -30,8 +24,8 @@ tap.test(`a message is logged and the operation is aborted
 tap.test(`a message is logged and the operation is aborted
         when the current directory is targeted using './'`, function ( t ) {
     const folder = './'
-    const throwing = function () { clearFolder( [folder] ) }
-    t.throws( throwing, folderErrObj( folder ) )
+    const returnValue = clearFolder( [folder] )
+    t.ok( returnValue === -1, `no operation` )
 
     t.end()
 })
@@ -39,8 +33,8 @@ tap.test(`a message is logged and the operation is aborted
 tap.test(`a message is logged and the operation is aborted
         when the parent directory is targeted using '..'`, function ( t ) {
     const folder = '..'
-    const throwing = function () { clearFolder( [folder] ) }
-    t.throws( throwing, folderErrObj( folder ) )
+    const returnValue = clearFolder( [folder] )
+    t.ok( returnValue === -1, `no operation` )
 
     t.end()
 })
@@ -48,8 +42,8 @@ tap.test(`a message is logged and the operation is aborted
 tap.test(`a message is logged and the operation is aborted
         when a sibling directory is targeted using '../*'`, function ( t ) {
     const folder = '../other-project'
-    const throwing = function () { clearFolder( [folder] ) }
-    t.throws( throwing, folderErrObj( folder ) )
+    const returnValue = clearFolder( [folder] )
+    t.ok( returnValue === -1, `no operation` )
 
     t.end()
 })